@@ -2,12 +2,13 @@
   <PropertyGroup>
     <Authors>Nicolas Musset</Authors>
     <Copyright>Copyright © Nicolas Musset 2016-2021</Copyright>
-    <Description></Description>
+    <Description>
+    </Description>
     <AssemblyName>Markdig.Wpf.SampleApp</AssemblyName>
     <AssemblyTitle>Markdig.Wpf.SampleApp</AssemblyTitle>
     <NeutralLanguage>en-US</NeutralLanguage>
     <VersionPrefix>0.5.0.1</VersionPrefix>
-    <TargetFrameworks>net452;netcoreapp3.1;net5.0-windows</TargetFrameworks>
+    <TargetFrameworks>net8.0-windows</TargetFrameworks>
     <ApplicationIcon />
     <OutputType>WinExe</OutputType>
     <StartupObject />
@@ -20,7 +21,16 @@
   <ItemGroup>
     <EmbeddedResource Update="Properties\Resources.resx" Generator="ResXFileCodeGenerator" LastGenOutput="Resources.Designer.cs" />
     <Compile Update="Properties\Resources.Designer.cs" DesignTime="True" AutoGen="True" DependentUpon="Resources.resx" />
+    <Compile Update="Properties\Settings.Designer.cs">
+      <DesignTimeSharedInput>True</DesignTimeSharedInput>
+      <AutoGen>True</AutoGen>
+      <DependentUpon>Settings.settings</DependentUpon>
+    </Compile>
     <Compile Update="Settings.Designer.cs" AutoGen="True" DependentUpon="Settings.settings" />
+    <None Update="Properties\Settings.settings">
+      <Generator>SettingsSingleFileGenerator</Generator>
+      <LastGenOutput>Settings.Designer.cs</LastGenOutput>
+    </None>
     <None Update="Settings.settings" LastGenOutput="Settings.Designer.cs" Generator="SettingsSingleFileGenerator" />
     <None Remove="Documents\.gitkeep" />
     <None Remove="img\.gitkeep" />
