@@ -7,7 +7,7 @@
     <AssemblyTitle>Markdig.Wpf.SampleAppCustomized</AssemblyTitle>
     <NeutralLanguage>en-US</NeutralLanguage>
     <VersionPrefix>0.5.0.1</VersionPrefix>
-    <TargetFrameworks>net452;netcoreapp3.1;net5.0-windows</TargetFrameworks>
+    <TargetFrameworks>net8.0-windows</TargetFrameworks>
     <ApplicationIcon />
     <OutputType>WinExe</OutputType>
     <StartupObject />
