@@ -1,13 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
   <PropertyGroup>
-    <Authors>Nicolas Musset</Authors>
-    <Copyright>Copyright © Nicolas Musset 2016-2021</Copyright>
-    <Description>A WPF library for lunet-io/markdig</Description>
+    <Authors>Nicolas Musset, Improved and customized by RYCB Studio for RYCBEditorX</Authors>
+    <Copyright>Copyright © Nicolas Musset 2016-2021, RYCB Studio 2024</Copyright>
+    <Description>A WPF library for lunet-io/markdig, which was customized for RYCB Editor X.</Description>
     <AssemblyName>Markdig.Wpf</AssemblyName>
     <AssemblyTitle>Markdig.Wpf</AssemblyTitle>
     <NeutralLanguage>en-US</NeutralLanguage>
     <VersionPrefix>0.5.0.1</VersionPrefix>
-    <TargetFrameworks>net452;netcoreapp3.1;net5.0-windows</TargetFrameworks>
+    <TargetFrameworks>net8.0-windows</TargetFrameworks>
     <SignAssembly Condition="'$(Configuration)' == 'Release_Signed'">true</SignAssembly>
     <PackageId>Markdig.Wpf</PackageId>
     <PackageId Condition="'$(SignAssembly)' == 'true'">Markdig.Wpf.Signed</PackageId>
